Fixed frequency markers and frequency change for CW SPLIT mode. Added scroll wheel to change split modes. Added scroll wheel to change Zoom. Added scroll wheel to chnage Step value. Added mouse drag for AF Gain slider and AGC Gain slider. CTUN and DUP for Receiver (was Radio). Added temperature display for SoapySDR devices (PlutoSDR and LimeSDR tested). Fixed discovery of PlutoSDR when connected by ethernet.

+++ transmitter.h
@@ -152,7 +152,7 @@ extern TRANSMITTER *create_transmitter(int channel);
 extern void transmitter_init_analyzer(TRANSMITTER *tx);
 extern void transmitter_save_state(TRANSMITTER *tx);
 extern void transmitter_restore_state(TRANSMITTER *tx);
-extern void add_mic_sample(TRANSMITTER *tx,short sample);
+extern void add_mic_sample(TRANSMITTER *tx,float sample);
 extern void transmitter_set_filter(TRANSMITTER *tx,int low,int high);
 extern void transmitter_set_pre_emphasize(TRANSMITTER *tx,int state);
 extern void transmitter_set_mode(TRANSMITTER *tx,int mode);
